+++ ChangeLog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Version [0.8.1](https://github.com/MC2IT/Agicap.ps1/compare/v0.8.0...v0.8.1)
+- **Treasury bank journal:** fixed the `Submit-BankJournalExport` cmdlet when the API returns no content.
+
 ## Version [0.8.0](https://github.com/MC2IT/Agicap.ps1/compare/v0.7.1...v0.8.0)
 - **Chart of accounts:** renamed the `-Number` parameter of the `Remove-AccountingAccount` cmdlet to `-AccountingAccountNumber`.
 - **Chart of accounts:** renamed the `-Code` parameter of the `Remove-ThirdParty` cmdlet to `-ThirdPartyCode`.